requirements.txt
tensorflow-macos>=2.8.0
tensorflow-metal>=0.5.0
scikit-learn>=0.24.0
numpy>=1.19.0
matplotlib>=3.4.0
seaborn>=0.11.0
imgaug>=0.4.0